Delayed Adstock Weights

delayed_adstock_weights is a function that is called to generate a vector of weights that define how media impressions will be decayed in present and carried over into the future.

Usage

delayed_adstock_weights(
  decay = 0.5,
  delay = 2,
  max_carryover = 12,
  normalize = TRUE
)

Arguments

decay

decay rate

delay

delay in the peak of the adstock

max_carryover

maximum carryover

normalize

should the result be scaled?

Value

a numeric vector of weights of length (max_carryover)

Examples

delayed_adstock_weights()
#>  [1] 2.421931e-01 4.843862e-01 2.421931e-01 3.027414e-02 9.460668e-04
#>  [6] 7.391147e-06 1.443583e-08 7.048747e-12 8.604427e-16 2.625863e-20
#> [11] 2.003374e-25 3.821133e-31
delayed_adstock_weights(decay = 0.1, delay = 2)
#>  [1]  8.332639e-02  8.332639e-01  8.332639e-02  8.332639e-05  8.332639e-10
#>  [6]  8.332639e-17  8.332639e-26  8.332639e-37  8.332639e-50  8.332639e-65
#> [11]  8.332639e-82 8.332639e-101
delayed_adstock_weights(decay = 0.1, delay = 5)
#>  [1] 8.331945e-17 8.331945e-10 8.331945e-05 8.331945e-02 8.331945e-01
#>  [6] 8.331945e-02 8.331945e-05 8.331945e-10 8.331945e-17 8.331945e-26
#> [11] 8.331945e-37 8.331945e-50