Weibull Adstock Weights

weibull_adstock_weights is a function that is called to generate a vector of weights that define how media impressions will be decayed in present and carried over into the future.

Usage

weibull_adstock_weights(
  shape = 1,
  scale = 1,
  max_carryover = 12,
  normalize = TRUE
)

Arguments

shape

weibull shape parameter

scale

weibull scale parameter

max_carryover

max_carryover parameter

normalize

should the result be scaled?

Value

a numeric vector of weights of length (max_carryover)

Examples

weibull_adstock_weights()
#>  [1] 6.321940e-01 2.325645e-01 8.554902e-02 3.146505e-02 1.156867e-02
#>  [6] 4.249202e-03 1.556519e-03 5.659370e-04 2.015221e-04 6.746135e-05
#> [11] 1.814315e-05 0.000000e+00
weibull_adstock_weights(shape = 0.5, scale = 10)
#>  [1] 0.353932732 0.199028228 0.132834001 0.094652032 0.069389599 0.051283565
#>  [7] 0.037604383 0.026875094 0.018220000 0.011083937 0.005096428 0.000000000
weibull_adstock_weights(shape = 1.5, scale = 20)
#>  [1] 0.00000000 0.03412033 0.05775276 0.07516472 0.08808964 0.09748554
#>  [7] 0.10398070 0.10803505 0.11001103 0.11020884 0.10888550 0.10626587